refactor: separate KVPP broadcast from KV pooling
